Equilibrage ennemis

See merge request PROJET_INFO3/gg/game-jam-2023!174

+++ game-project/game/info3/game/modele/MoveableEntity.java
@@ -64,7 +64,7 @@ public abstract class MoveableEntity extends Entity {
 		return m_damage;
 	}
 
-	public void setDamage(float m_damage) {
+	public void setDamage(int m_damage) {
 		this.m_damage = m_damage;
 	}
